requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
pymysql==1.1.1
python-multipart==0.0.9
jinja2==3.1.4
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.8.2

